requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-restful
flask-security-too
flask-cors
redis
email_validator
bcrypt
celery[redis]